// src/app/shared/models/profesor.ts
export interface Profesor{
    idProfesor?:            number;
    nombre?:                string;
    apellido?:              string
    correo?:                string;
    foto?:                  string;
    usuarioCreador?:        string;
    usuarioModificador?:    string;
    fechaCreacion?:         Date;
    fechaModificacion?:     Date;
    idGenero?:              number
}